Fix bug where dialpad does not get dismissed.
There was a delay when updating the current fragment to be the speed
dial fragment. Removing this delay allows the fragment to be updated
immediately and the dialpad to be shown.

The clean state from this ensures that the dialpad is properly
dismissed.

Lastly, clean up some code in TelecomActivity.

Test: manually tested
Bug: 32380006
Change-Id: Ice8c7b4f5db8cb2fe704c74a66de30c8d99c5e5e