Make the Dialer direct boot aware.
The Dialer needs this attribute because it might be possible for a user
to already be in a phone call when AAE boots up. The Dialer needs to be
notified of this state when it boots so that it can reflect the phone
call status properly.

Test: make a phone call, reboot AAE and ensure the Dialer shows the
in-call screen when AAE is booted up.
Bug: 64073836

Change-Id: I985bb3460b8e3d8115c90fcb655cd8be0a3006e0