Add long-press functionality to the "0", "1", and delete dialpad keys.
Closely mimics the behavior of the phone dialer app. Each key tone now
plays as long as the touch is held down, and stops when the touch is
released.  Long-pressing on 0 initially starts with 0, then replaces it
with '+', and stops the key tone.  Long-press on 1 attempts to call the
voicemail number if it's available from the OS.  Long-press on delete
clears the current number.

Change-Id: Icf3675b5aab89c4dacf93d967f6c15e22e0dbc7e
Bug: 73481550
Fixes: 73481550
Test: Manually verified on Mojave.