Support incoming call on top of user-switcher(keyguard)
Add use-permission for WAKE_LOCK and related flags in TelecomActivity,
so that InCallUI can be shown on top of keyguard upon incoming call.

Here is the tech details to explain why the related flags are also configured
in onCreate() besides onCallAdded() and onCallRemoved():
When keyguard is locked, upon incoming call, TelecomActivity will be launched
the system will set TelecomAcitity to be invisible below keyguard and try
to move it to STOPPED state, so there is a chance (the race condition) that
onStop() will be called before onCallAdded(), in which, mUiCallManager will
remove listener for TelecomActivity, then onCallAdded() won't be called.
So we need to configure the related flags in onCreate() too.

Test: pair phone with DUT, lock the screen, make a call to the phone

bug: 121155028

Change-Id: Ic50e598d99c12b001d76edd1ef49b59c864936f4
Signed-off-by: robinz1x <robinx.zhang@intel.com>
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>