UI fix for contact details page.
Added touch feedback for phone number list item.
Each list item view is wrapped in a ColumnCarView
Set the recycler view to wrap content on width and center in parent.
Use the VerticalListDividerDecoration for dividers.
Create outline providers for top/bottom corners so the entire list looks
like a card.

Bug: 121278955
Fixes: 121278955
Test: manual testing
Change-Id: Iddcce6c949534bf689f9a2255357d721871e8863