Incoming call notificaiton with answer and decline actions.
If state is ringing, show the HUN no matter which page user is on.
Tapping on the notification will bring user to the InCall page and
dimiss the notification.
If user was outside of Dialer and open dialer, user will go to the
incall page and notification will not dismiss until call gets answered
or declined(No-op since open incall page is temporary).
If user was on the incall page, receiving an incoming call will both
have the HUN and the incoming call page(WIP for HUN or incoming call
page only for this case).

Test: manual testing
Bug: 120173390

Change-Id: I8009daa85a65e4694b5fd542944a7374ba13ab66