Make small adjustments on Portrait Mode Dialpad
The previous port-mode dialpad is to chain title, keypad, and call
button together and put the whole thing together in the middle of
dialpad page
And the new version is to put call button at a fixed position--the same
position as onGoingCallControllerBar, chain title and keypad
together, and put them in the center of --
1. top of parent and call button in dialpad page
2. top of parent and onGoingCallControllerBar in on going call page

Also limited the input number to be a limited length, which is 20,
for now--
If the input number is equal or smaller than 20, show every digit.
If the input number is longer than 20, show the last 20 digits.

Bug: 128455180
Test: Manually
Change-Id: Ie4dff4a5416a225c9d7ceff5f17a437fceba18f5