Fix weird call duration when system time rewinds
We used to be using android.text.format.DateUtils.formatElapsedTime()
to format the call duration, but it doesn't support negatives, and
was giving us strings like "00:-136". Phone's Dialer uses Chronometer,
which does handle negative time correctly, so I switched to that.

Changing the system time will still cause the call duration to change
and potentially go into the negatives, but this behavior matches
what happens on the phone.

Bug: 133146985
Test: Manually
Change-Id: If2ad34cfab813c5b70611a46d2e3a436672a9548