Stop whole screen redrawing every second in ongoing call
TextView calls requestLayout() whenever its text is changed and the
view doesn't have a fixed width. requestLayout() will propagate to
the root node and cause the whole screen to be redrawn. In our
ongoing call page we update a TextView (Chronometer) every second,
so the whole screen is being redrawn every second.

This CL changes the Chronometers' widths to match_parent to fix that.

You can see the effects of this with the "Show GPU view updates"
developer option.

Bug: None
Test: Manually; Show GPU View updates developer option
Change-Id: If78bbf326f8625ea4cf098611db2a434d7fa17a4