Remember scroll position in lists
If a recyclerview briefly displays an empty list, it will reset
it's scroll position to the top of the list, even when items are added
later. This CL makes some tweaks to prevent our call history and
contact lists from being emptied and reloaded whenever you navigate
away from them and back, so that the scroll bar position isn't lost.

Bug: 135700452
Test: Manually
Change-Id: I1cf85d8c181111eee738b1140c1bbe490fe6c879