Support favorite bit from phone.
After the change:
Bluetooth favorite contacts show up at the head of list.
Local favorite contacts show up at the end of list.
Bluetooth favorite entries will be disabled in the add-to-favorite flow
and in the contact detail page with favorite indicator on.
Right now there is no indicator for the source of the favorite contacts
if it is from phone or it is local.

Test: manually
Bug: 143498802
Change-Id: I97c6e6a914f9408e339bceb9835760dde962dde3