Fix no display name for saved contacts for HUN.
This issue is basically a race condition that when the callback is
triggerred, NotificationManager.getActiveNotifications() is not
returning the previous posted HUN which might turn active later. By
caching the notifications on our own, the check if the HUN exists will
be valid. We will cancel the notification explicitly on any action that
will dismiss the notification(action button and fullscreen intent)

Fixes: 146165572
Test: manually
Change-Id: I896dc75dff915f52702dd576ac30482c6c5c5788