Update Contacts icon
Previously ic_contact.xml was being shadowed by ic_contact.png.
Now that ic_contact is being deleted (ag/10320076), change
the xml version to be the same as the png version.

The png version / the new xml version is a simple contact person.
The old xml version was the contact person with a solid background.

Bug: 149503007
Test: Manually
Change-Id: I293a5e13f2a5927bca0b742b95d8048de5cc3cdd