Add the option to disable the contact details header
An OEM is trying to remove the contact details header by
making all the views in it have visibility GONE, but
there is currently a bug in recyclerview (b/165334939)
where doing that causes the scrollbar to not go all the
way up to the top of the screen.

As a workaround, add a boolean to disable the header.

Fixes: 164160249
Test: Manually
Change-Id: Ifb51aa9282f2346c074a3d7ecb6ac327d8b5e13e