Fix data loading in recents and contacts page
Handle the case when cursor is null and null data is set from
AsyncQueryLiveData. Before the fix, there will be no update triggered
from ViewModel which is loading state. After the fix, it will show empty
state.
Fix the issue that when turning bluetooth off or turning profile
permission off in bluetooth settings page, contacts and call logs are
not cleared.

Bug: 166323672
Bug: 167158164
Test: manual

Change-Id: I168dbf0a9da84d08791d2493f2821bd594132d27